Derive status builder
