+++ examples/upload_photo.rs
@@ -17,9 +17,9 @@ async fn run() -> Result<()> {
     };
 
     let media = mastodon.media(input, description).await?;
-    let status = StatusBuilder::new()
+    let status = StatusBuilder::default()
         .status("Mastodon-async photo upload example/demo (automated post)")
-        .media_ids([media.id])
+        .media_ids(vec![media.id])
         .visibility(Visibility::Private)
         .build()?;
     let status = mastodon.new_status(status).await?;
